adopt PEP 695 type syntax + refactor docstrings
refactored the singleton decorator implementation to use the new type parameter syntax + updated docstrings.
