--- singleton_decorator/decorator.py
@@ -1,32 +1,0 @@
-from typing import TypeVar, Any
-
-class _SingletonWrapper:
-    """
-    A singleton wrapper class. Its instances would be created
-    for each decorated class. 
-    """
-
-    def __init__(self, cls):
-        self.__wrapped__ = cls
-        self._instance = None
-
-    def __call__(self, *args, **kwargs):
-        """Returns a single instance of decorated class"""
-        if self._instance is None:
-            self._instance = self.__wrapped__(*args, **kwargs)
-        return self._instance
-
-    def __getattr__(self, name: str) -> Any:
-        """Allows class member access from the wrapper"""
-        return getattr(self.__wrapped__, name)
-
-# ensures a correct type hinting
-SingletonType = TypeVar("SingletonType")
-
-def singleton(cls: SingletonType) -> SingletonType:
-    """
-    A singleton decorator. Returns a wrapper objects. A call on that object
-    returns a single instance object of decorated class. Use the __wrapped__
-    attribute to access decorated class directly in unit tests
-    """
-    return _SingletonWrapper(cls) # type: ignore
